refactor(web): extract reports cards for reuse
